% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leslie_disaggregate.R
\name{GCD}
\alias{GCD}
\title{Greatest common divisor of a set of integers}
\usage{
GCD(x)
}
\arguments{
\item{x}{A numeric vector of positive integers.}
}
\value{
A positive integer giving the GCD of \code{x}.
}
\description{
Computes the greatest common divisor (GCD) of a numeric vector \code{x} of positive integers.
Used by \code{\link{LCM}}.
}
\keyword{internal}
