% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leslie_disaggregate.R
\name{LCM}
\alias{LCM}
\title{Least common multiple of two integers}
\usage{
LCM(a, b)
}
\arguments{
\item{a}{A positive integer.}

\item{b}{A positive integer.}
}
\value{
A positive integer giving the LCM of \code{a} and \code{b}.
}
\description{
Computes the least common multiple (LCM) of two positive integers.
Used by \code{\link{leslie_disaggregate}}.
}
\keyword{internal}
