% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{leslie_dominant_eigen}
\alias{leslie_dominant_eigen}
\title{Dominant eigen-elements of a Leslie matrix}
\usage{
leslie_dominant_eigen(A)
}
\arguments{
\item{A}{A Leslie matrix (checked with \code{\link{is_leslie}}). A length-1 positive numeric
will be coerced to a 1x1 matrix.}
}
\value{
A named list with elements:
\itemize{
  \item \code{lambda}: spectral radius of \code{A}
  \item \code{w}: stable age distribution, \code{sum(w)=1}
  \item \code{v}: reproductive values, \code{sum(v*w)=1}
}
}
\description{
Returns the dominant eigenvalue (\eqn{\lambda}) and the stable age distribution
\code{w} and reproductive values \code{v} for a Leslie matrix \code{A}.
}
\details{
Normalization:
The output is normalized using the conventions in \code{\link{leslie_stable_age}} and
\code{\link{leslie_reproductive_values}}
\itemize{
  \item \code{w} is scaled so that \code{sum(w) = 1}
  \item \code{v} is scaled so that \code{sum(v * w) = 1}
}
}
\examples{
leslie_dominant_eigen(1)

L <- matrix(c(
  0, 2, 1,
  0.6, 0, 0,
  0, 0.7, 0
), nrow = 3, byrow = TRUE)
dom <- leslie_dominant_eigen(L)
dom$lambda
sum(dom$w)
sum(dom$v * dom$w)

}
\references{
Caswell, H. (2001). \emph{Matrix population models: construction, analysis and
interpretation} (2nd ed.). Sinauer.

Demetrius, L. (1974). Demographic parameters and natural selection.
\emph{Proceedings of the National Academy of Sciences}, 71(12), 4645-4647.
\doi{10.1073/pnas.71.12.4645}
}
\seealso{
\code{\link{spectral_radius}}, \code{\link{leslie_stable_age}},
\code{\link{leslie_reproductive_values}}
}
